--- a/addons/mail/models/mail_thread.py
+++ b/addons/mail/models/mail_thread.py
@@ -3224,9 +3224,9 @@ class MailThread(models.AbstractModel):
         if not recipients_data:
             return recipients_data
         # cache data fetched by manual query to avoid extra queries when reading user.partner_id
-        for r in filter(lambda r: r["uid"], recipients_data):
-            user = self.env["res.users"].browse(r["uid"])
-            self.env.cache.insert_missing(user, user._fields["partner_id"], [r["id"]])
+        uid2pid = {r['uid']: r['id'] for r in recipients_data if r['uid']}
+        users = self.env['res.users'].browse(uid2pid)
+        users._fields['partner_id']._insert_cache(users, uid2pid.values())
         # if scheduled for later: add in queue instead of generating notifications
         scheduled_date = self._is_notification_scheduled(kwargs.pop('scheduled_date', None))
         if scheduled_date: