--- a/odoo/orm/fields_misc.py
+++ b/odoo/orm/fields_misc.py
@@ -28,11 +28,6 @@ class Boolean(Field[bool]):
     def convert_to_column(self, value, record, values=None, validate=True):
         return bool(value)
 
-    def convert_to_column_update(self, value, record):
-        if self.company_dependent:
-            value = {k: bool(v) for k, v in value.items()}
-        return super().convert_to_column_update(value, record)
-
     def convert_to_cache(self, value, record, validate=True):
         return bool(value)