… odoo/addons/base/tests/test_api.py
… odoo/addons/test_new_api/tests/test_new_fields.py
… odoo/orm/environments.py
… odoo/orm/fields_binary.py
… odoo/orm/fields_relational.py
… odoo/orm/fields_temporal.py
… odoo/orm/fields_textual.py
… odoo/orm/models.py