Moving the cache management to fields so that they can independently customize the behaviour over how the cache is structured.

The data of the transaction is now available on the `Transaction` object. The `Cache` object stays, but most of its low-level methods are deprecated and it is just manipulating the data of the transaction through the fields.
The `Environment` gains a few lazy properties to be able to quickly access the cache while doing a minimum amount of calls.
`Field` has now generic methods to manipulate the cache and sub-classes can overwrite them.

Some performance improvements (some operations are similar in time). Results are quite volatile, but most common operations are 5-25% faster).
| operation | before | after |
| --------| --| --|
| read partner.name | 450ns | 330ns |
| read user.partner_id | 660ns | 570ns |
| read 3 users.partner_id | 1.7us | 1.2us |
| read 3 users.bank_ids | 2.9us | 2.4us |

task-4139371

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr