@robodoo 2025-04-04
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/204796.png)](https://mergebot.odoo.com/odoo/odoo/pull/204796)