@kmagusiak 2025-04-08
> This is overall a very good job!
> 
> One thing I don't like is the naming convention (all cache-handling methods start with `_cache_`). It slightly optimizes those methods' definition, but makes the calls to those methods less readable: `field._cache_invalidate()` vs `field._invalidate_cache()`. The second one looks better to me.
> 
> As I pointed out below, I think we should get rid of `field._cache_view` and rely on the other one. That's what you do in most cases, anyway, when you use its cached version.
> 
> I also suggest to make the cache access API on the field, instead of on the environment (even if the cache of that method is on the environment.) So instead of having `env._field_cache(field)` being a cache to `field._cache_view_for_update(env)`, I suggest to have `field._get_cache(env)` to be a cache to `field._get_cache_impl(env)`.

I did most of the adaptations. Tomorrow, we can review them. The only part left is naming of some functions.

I like the `field._get_cache(env)`. Note that we still need a function that does not create entries (for debugging, flushing, and some other use cases including backward compatibility - otherwise, we may remove `Cache` object entirely in a future PR).