@rco-odoo 2025-05-07
@robodoo override=ci/security

* `type(val)` is used in code that has been reindented; nothing new here.
* `__repr__` of `LangProxyDict` is surrounded by `<>` to prevent any evaluation of the string.